pyproject.toml
[project]
name = "furniverse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastembed>=0.7.4",
    "numpy>=2.4.1",
    "pillow>=10.3.0,<12.0.0",
    "qdrant-client>=1.16.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
]
